# backend/openapi.yaml
openapi: 3.0.0
info:
  title: NGO Impact Tracker API
  version: 1.0.0
  description: |
    API for tracking NGO impact reports across India.
    Supports individual report submission, bulk CSV uploads with background processing,
    and aggregated dashboard analytics.
  contact:
    name: WeDoGood
servers:
  - url: http://localhost:5000
    description: Local development server

paths:
  /health:
    get:
      summary: Health check
      tags: [System]
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: string
                    format: date-time

  /report:
    post:
      summary: Submit a single report
      tags: [Reports]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportInput'
      responses:
        '201':
          description: Report created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /reports/upload:
    post:
      summary: Upload CSV file for bulk processing
      tags: [Reports]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV file with reports
      responses:
        '202':
          description: Upload accepted, processing started
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  job_id:
                    type: string
                    format: uuid
                  filename:
                    type: string
        '400':
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /job-status/{job_id}:
    get:
      summary: Get job processing status
      tags: [Jobs]
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
        '404':
          description: Job not found

  /job-status:
    get:
      summary: List all jobs
      tags: [Jobs]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, completed, failed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/JobStatus'

  /dashboard:
    get:
      summary: Get aggregated dashboard data for a month
      tags: [Dashboard]
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
          example: '2024-03'
      responses:
        '200':
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardData'
        '400':
          description: Invalid month format

  /dashboard/months:
    get:
      summary: Get list of months with data
      tags: [Dashboard]
      responses:
        '200':
          description: List of months
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: string
                      example: '2024-03'

  /dashboard/trends:
    get:
      summary: Get monthly trends
      tags: [Dashboard]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        '200':
          description: Monthly trends data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MonthlyTrend'

  /dashboard/reports:
    get:
      summary: Get paginated reports with filtering
      tags: [Dashboard]
      parameters:
        - name: month
          in: query
          schema:
            type: string
        - name: ngo_id
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated reports
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedReports'

  /dashboard/ngos:
    get:
      summary: Get all NGOs with stats
      tags: [Dashboard]
      responses:
        '200':
          description: List of NGOs
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/NGOSummary'

components:
  schemas:
    ReportInput:
      type: object
      required:
        - ngo_id
        - month
        - people_helped
        - events_conducted
        - funds_utilized
      properties:
        ngo_id:
          type: string
          example: NGO-001
        month:
          type: string
          pattern: '^\d{4}-\d{2}$'
          example: '2024-03'
        people_helped:
          type: integer
          minimum: 0
          example: 150
        events_conducted:
          type: integer
          minimum: 0
          example: 5
        funds_utilized:
          type: number
          minimum: 0
          example: 50000

    ReportResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            id:
              type: integer
            ngo_id:
              type: string
            month:
              type: string
            people_helped:
              type: integer
            events_conducted:
              type: integer
            funds_utilized:
              type: number

    JobStatus:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        filename:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        total_rows:
          type: integer
        processed_rows:
          type: integer
        successful_rows:
          type: integer
        failed_rows:
          type: integer
        errors:
          type: array
          items:
            type: object
            properties:
              row:
                type: integer
              error:
                type: string
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true

    DashboardData:
      type: object
      properties:
        success:
          type: boolean
        month:
          type: string
        data:
          type: object
          properties:
            total_ngos_reporting:
              type: integer
            total_people_helped:
              type: integer
            total_events_conducted:
              type: integer
            total_funds_utilized:
              type: number

    MonthlyTrend:
      type: object
      properties:
        month:
          type: string
        total_ngos_reporting:
          type: integer
        total_people_helped:
          type: integer
        total_events_conducted:
          type: integer
        total_funds_utilized:
          type: number

    NGOSummary:
      type: object
      properties:
        ngo_id:
          type: string
        reports_count:
          type: integer
        total_people_helped:
          type: integer
        total_events_conducted:
          type: integer
        total_funds_utilized:
          type: number

    PaginatedReports:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/ReportInput'
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            totalCount:
              type: integer
            totalPages:
              type: integer
            hasNext:
              type: boolean
            hasPrev:
              type: boolean

    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        details:
          type: array
          items:
            type: string